readyup|refactor: Split kit-source resolution out of the run command
Extracts kit-source resolution from `readyup`'s `src/run/runCommand.ts` into five modules named for what they export: `resolveKitSources.ts` holds the branch that reads the source flags, `resolveFromSource.ts` the `--from` switch and its URL builders, `resolveConfiguredPackages.ts` the `--packages` expansion, and `ResolvedKitEntry.ts` and `defaultKitName.ts` the two symbols both sides of that edge read. `runCommand.ts` drops from 810 lines to 536 and now imports the entry type rather than declaring it.

`describeUrlSource` moves with resolution rather than execution, its only caller being `resolveKitSources`. `resolveInstalledPackageRoot` goes with `resolveFromSource`, whose `npm:` case is its only caller.

Each module's tests call it directly and mock nothing, so the `--from` cases construct a parsed source instead of routing a flag string through `parseFromValue`, which has its own coverage. `resolveConfiguredPackages` gains a test file of its own, the `--packages` branch having had none outside the wiring test. Branch coverage of `src/run` rises from 95.71% to 96.63%, with all five new modules fully covered.
