+++ packages/readyup/src/bin/route.ts
@@ -18,7 +18,8 @@ import { listCommand } from '../list/listCommand.ts';
 import { writeHuman } from '../output/writeHuman.ts';
 import { formatJsonError } from '../reporting/formatJsonError.ts';
 import { parseRunArgs } from '../run/parseRunArgs.ts';
-import { resolveKitSources, runCommand } from '../run/runCommand.ts';
+import { resolveKitSources } from '../run/resolveKitSources.ts';
+import { runCommand } from '../run/runCommand.ts';
 import { verifyCommand } from '../verify/verifyCommand.ts';
 import { VERSION } from '../version.ts';
 import { EXIT_OK, EXIT_TOOL_FAILURE } from './exitCodes.ts';
