+++ packages/readyup/src/run/__tests__/runCommand.unit.test.ts
@@ -1,5 +1,4 @@
 import assert from 'node:assert';
-import path from 'node:path';
 import process from 'node:process';
 
 import { afterEach, beforeEach, describe, expect, it, type MockInstance, vi } from 'vitest';
@@ -58,477 +57,7 @@ vi.mock(import('../../remote/loadRemoteKit.ts'), () => ({
   loadRemoteKit: mockLoadRemoteKit,
 }));
 
-import packageJson from '../../../package.json' with { type: 'json' };
-import { resolveKitSources, runCommand } from '../runCommand.ts';
-
-/** Repo root, which is where the workspace link for `readyup` lives. */
-const REPO_ROOT = path.resolve(import.meta.dirname, '../../../../..');
-
-describe(resolveKitSources, () => {
-  /** Build args with defaults for internal config. */
-  function resolve(
-    overrides: Partial<Parameters<typeof resolveKitSources>[0]> = {},
-  ): ReturnType<typeof resolveKitSources> {
-    return resolveKitSources({
-      filePath: undefined,
-      fromValue: undefined,
-      urlValue: undefined,
-      kitSpecifiers: [],
-      checklists: undefined,
-      jit: false,
-      internal: false,
-      internalDir: '.',
-      internalInfix: undefined,
-      ...overrides,
-    });
-  }
-
-  // -- Default resolution (compiled .js) --
-
-  it('resolves default kit path to .js', () => {
-    expect(resolve()).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/default.js' }, checklists: [] },
-    ]);
-  });
-
-  it('resolves named kit from positional specifier', () => {
-    expect(resolve({ kitSpecifiers: [{ kitName: 'deploy', checklists: [] }] })).toStrictEqual([
-      { name: 'deploy', source: { path: '.readyup/kits/deploy.js' }, checklists: [] },
-    ]);
-  });
-
-  it('resolves slash-separated kit name', () => {
-    expect(resolve({ kitSpecifiers: [{ kitName: 'shared/deploy', checklists: [] }] })).toStrictEqual([
-      { name: 'shared/deploy', source: { path: '.readyup/kits/shared/deploy.js' }, checklists: [] },
-    ]);
-  });
-
-  it('applies --checklists to the named kit', () => {
-    expect(
-      resolve({ kitSpecifiers: [{ kitName: 'deploy', checklists: [] }], checklists: ['build', 'test'] }),
-    ).toStrictEqual([{ name: 'deploy', source: { path: '.readyup/kits/deploy.js' }, checklists: ['build', 'test'] }]);
-  });
-
-  it('applies --checklists to the default kit when no kit is named', () => {
-    expect(resolve({ checklists: ['build'] })).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/default.js' }, checklists: ['build'] },
-    ]);
-  });
-
-  // -- npm: source --
-
-  // `readyup` is linked into this repo's node_modules as a workspace package, so it stands in for any
-  // installed dependency without needing a fixture.
-  it('resolves a kit inside an installed package', () => {
-    const [entry] = resolve({ fromValue: 'npm:readyup' });
-
-    expect(entry?.name).toBe('default');
-    expect(entry?.source).toStrictEqual({
-      path: path.join(REPO_ROOT, 'packages', 'readyup', '.readyup', 'kits', 'default.js'),
-    });
-  });
-
-  it('resolves a named kit inside an installed package', () => {
-    const [entry] = resolve({ fromValue: 'npm:readyup', kitSpecifiers: [{ kitName: 'drift', checklists: [] }] });
-
-    expect(entry?.source).toStrictEqual({
-      path: path.join(REPO_ROOT, 'packages', 'readyup', '.readyup', 'kits', 'drift.js'),
-    });
-  });
-
-  // The provenance is what names the copy a check ran against, which is the whole point of resolving from an
-  // installed package. It reads from the same manifest the resolver reads, so a version bump leaves this alone.
-  it('carries the package and its installed version as the kit provenance', () => {
-    const [entry] = resolve({ fromValue: 'npm:readyup' });
-
-    expect(entry?.provenance).toStrictEqual({
-      kind: 'package',
-      packageName: 'readyup',
-      version: packageJson.version,
-    });
-  });
-
-  it('rejects a version spec by naming the flag that reaches a published kit', () => {
-    expect(() => resolve({ fromValue: 'npm:readyup@0.22.0' })).toThrow(/not supported yet[\s\S]*--url/);
-  });
-
-  it('rejects an uninstalled package by naming the direct-dependency requirement', () => {
-    expect(() => resolve({ fromValue: 'npm:readyup-package-that-does-not-exist' })).toThrow(
-      /is not installed; it must be a direct dependency/,
-    );
-  });
-
-  // -- --jit flag --
-
-  it('resolves to .ts with --jit', () => {
-    expect(resolve({ jit: true })).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/default.ts' }, checklists: [] },
-    ]);
-  });
-
-  // -- --internal flag --
-
-  it('applies internal dir with --internal', () => {
-    expect(resolve({ internal: true, internalDir: 'internal' })).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/internal/default.js' }, checklists: [] },
-    ]);
-  });
-
-  it('applies internal dir and infix with --internal', () => {
-    expect(resolve({ internal: true, internalDir: 'internal', internalInfix: 'int' })).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/internal/default.int.js' }, checklists: [] },
-    ]);
-  });
-
-  it('combines --jit and --internal', () => {
-    expect(resolve({ jit: true, internal: true, internalDir: 'internal', internalInfix: 'int' })).toStrictEqual([
-      { name: 'default', source: { path: '.readyup/kits/internal/default.int.ts' }, checklists: [] },
-    ]);
-  });
-
-  it('applies internal dir with named kit', () => {
-    expect(
-      resolve({
-        kitSpecifiers: [{ kitName: 'deploy', checklists: [] }],
-        internal: true,
-        internalDir: 'internal',
-        internalInfix: 'int',
-      }),
-    ).toStrictEqual([{ name: 'deploy', source: { path: '.readyup/kits/internal/deploy.int.js' }, checklists: [] }]);
-  });
-
-  // -- External sources without config fields --
-
-  it('resolves --file without internalDir/internalInfix', () => {
-    expect(
-      resolveKitSources({
-        filePath: 'custom/path.ts',
-        fromValue: undefined,
-        urlValue: undefined,
-        kitSpecifiers: [],
-        checklists: undefined,
-        jit: false,
-        internal: false,
-      }),
-    ).toStrictEqual([
-      {
-        name: 'path',
-        source: { path: 'custom/path.ts' },
-        checklists: [],
-        provenance: { kind: 'directory', label: 'custom' },
-      },
-    ]);
-  });
-
-  it('resolves --url without internalDir/internalInfix', () => {
-    expect(
-      resolveKitSources({
-        filePath: undefined,
-        fromValue: undefined,
-        urlValue: 'https://example.com/kit.js',
-        kitSpecifiers: [],
-        checklists: undefined,
-        jit: false,
-        internal: false,
-      }),
-    ).toStrictEqual([
-      {
-        name: 'kit',
-        source: { url: 'https://example.com/kit.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'example.com/kit.js' },
-      },
-    ]);
-  });
-
-  it('resolves --from without internalDir/internalInfix', () => {
-    expect(
-      resolveKitSources({
-        filePath: undefined,
-        fromValue: 'github:org/repo',
-        urlValue: undefined,
-        kitSpecifiers: [{ kitName: 'deploy', checklists: [] }],
-        checklists: undefined,
-        jit: false,
-        internal: false,
-      }),
-    ).toStrictEqual([
-      {
-        name: 'deploy',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/deploy.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-    ]);
-  });
-
-  // -- --file flag --
-
-  it('resolves --file to a single path source entry', () => {
-    expect(resolve({ filePath: 'custom/path.ts' })).toStrictEqual([
-      {
-        name: 'path',
-        source: { path: 'custom/path.ts' },
-        checklists: [],
-        provenance: { kind: 'directory', label: 'custom' },
-      },
-    ]);
-  });
-
-  it('resolves --file with --checklists', () => {
-    expect(resolve({ filePath: 'custom/path.ts', checklists: ['c1', 'c2'] })).toStrictEqual([
-      {
-        name: 'path',
-        source: { path: 'custom/path.ts' },
-        checklists: ['c1', 'c2'],
-        provenance: { kind: 'directory', label: 'custom' },
-      },
-    ]);
-  });
-
-  // -- --from github: --
-
-  it('resolves --from github: without ref to a URL with main ref', () => {
-    expect(
-      resolve({ fromValue: 'github:org/repo', kitSpecifiers: [{ kitName: 'nmr', checklists: [] }] }),
-    ).toStrictEqual([
-      {
-        name: 'nmr',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/nmr.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-    ]);
-  });
-
-  it('resolves --from github: with ref', () => {
-    expect(
-      resolve({ fromValue: 'github:org/repo@v1', kitSpecifiers: [{ kitName: 'nmr', checklists: [] }] }),
-    ).toStrictEqual([
-      {
-        name: 'nmr',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/v1/.readyup/kits/nmr.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@v1' },
-      },
-    ]);
-  });
-
-  it('defaults --from github: kit to "default"', () => {
-    expect(resolve({ fromValue: 'github:org/repo' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/default.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-    ]);
-  });
-
-  it('resolves multiple kits with --from github:', () => {
-    expect(
-      resolve({
-        fromValue: 'github:org/repo',
-        kitSpecifiers: [
-          { kitName: 'deploy', checklists: [] },
-          { kitName: 'infra', checklists: ['c1'] },
-        ],
-      }),
-    ).toStrictEqual([
-      {
-        name: 'deploy',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/deploy.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-      {
-        name: 'infra',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/infra.js' },
-        checklists: ['c1'],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-    ]);
-  });
-
-  // -- --from bitbucket: --
-
-  it('resolves --from bitbucket: to a Bitbucket Cloud API source URL', () => {
-    expect(
-      resolve({ fromValue: 'bitbucket:myteam/deploy-checks', kitSpecifiers: [{ kitName: 'deploy', checklists: [] }] }),
-    ).toStrictEqual([
-      {
-        name: 'deploy',
-        source: {
-          url: 'https://api.bitbucket.org/2.0/repositories/myteam/deploy-checks/src/main/.readyup/kits/deploy.js',
-        },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'bitbucket:myteam/deploy-checks@main' },
-      },
-    ]);
-  });
-
-  it('resolves --from bitbucket: with ref', () => {
-    expect(resolve({ fromValue: 'bitbucket:myteam/repo@v2' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { url: 'https://api.bitbucket.org/2.0/repositories/myteam/repo/src/v2/.readyup/kits/default.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'bitbucket:myteam/repo@v2' },
-      },
-    ]);
-  });
-
-  // -- --from local path --
-
-  it('resolves --from with local path to a .js path under .readyup/kits/', () => {
-    expect(resolve({ fromValue: '/path/to/repo' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { path: '/path/to/repo/.readyup/kits/default.js' },
-        checklists: [],
-        provenance: { kind: 'directory', label: '/path/to/repo/.readyup/kits' },
-      },
-    ]);
-  });
-
-  it('resolves --from with relative local path against cwd', () => {
-    const expected = path.resolve(process.cwd(), '../sibling-repo');
-
-    expect(resolve({ fromValue: '../sibling-repo' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { path: `${expected}/.readyup/kits/default.js` },
-        checklists: [],
-        provenance: { kind: 'directory', label: '../sibling-repo/.readyup/kits' },
-      },
-    ]);
-  });
-
-  it('resolves multiple kits with --from local path', () => {
-    expect(
-      resolve({
-        fromValue: '/path/to/repo',
-        kitSpecifiers: [
-          { kitName: 'deploy', checklists: [] },
-          { kitName: 'infra', checklists: [] },
-        ],
-      }),
-    ).toStrictEqual([
-      {
-        name: 'deploy',
-        source: { path: '/path/to/repo/.readyup/kits/deploy.js' },
-        checklists: [],
-        provenance: { kind: 'directory', label: '/path/to/repo/.readyup/kits' },
-      },
-      {
-        name: 'infra',
-        source: { path: '/path/to/repo/.readyup/kits/infra.js' },
-        checklists: [],
-        provenance: { kind: 'directory', label: '/path/to/repo/.readyup/kits' },
-      },
-    ]);
-  });
-
-  // -- --from global --
-
-  it('resolves --from global to home directory', () => {
-    const homeDir = process.env['HOME'] ?? process.env['USERPROFILE'] ?? '~';
-
-    expect(resolve({ fromValue: 'global' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { path: `${homeDir}/.readyup/kits/default.js` },
-        checklists: [],
-        provenance: { kind: 'directory', label: '~/.readyup/kits' },
-      },
-    ]);
-  });
-
-  // -- --from dir: --
-
-  it('resolves --from dir: to an arbitrary directory', () => {
-    const resolved = path.resolve(process.cwd(), 'custom/kits');
-
-    expect(resolve({ fromValue: 'dir:custom/kits' })).toStrictEqual([
-      {
-        name: 'default',
-        source: { path: `${resolved}/default.js` },
-        checklists: [],
-        provenance: { kind: 'directory', label: 'custom/kits' },
-      },
-    ]);
-  });
-
-  // -- --url flag --
-
-  it('resolves --url to a URL source', () => {
-    expect(resolve({ urlValue: 'https://example.com/config.js' })).toStrictEqual([
-      {
-        name: 'config',
-        source: { url: 'https://example.com/config.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'example.com/config.js' },
-      },
-    ]);
-  });
-
-  it('resolves --url with --checklists', () => {
-    expect(resolve({ urlValue: 'https://example.com/config.js', checklists: ['c1', 'c2'] })).toStrictEqual([
-      {
-        name: 'config',
-        source: { url: 'https://example.com/config.js' },
-        checklists: ['c1', 'c2'],
-        provenance: { kind: 'remote', label: 'example.com/config.js' },
-      },
-    ]);
-  });
-
-  // -- Isolation of internal config with source flags --
-
-  it('ignores internal config when --file is used', () => {
-    expect(
-      resolve({ filePath: 'custom/path.ts', internal: true, internalDir: 'internal', internalInfix: 'int' }),
-    ).toStrictEqual([
-      {
-        name: 'path',
-        source: { path: 'custom/path.ts' },
-        checklists: [],
-        provenance: { kind: 'directory', label: 'custom' },
-      },
-    ]);
-  });
-
-  it('ignores internal config when --from is used', () => {
-    expect(
-      resolve({ fromValue: 'github:org/repo', internal: false, internalDir: 'internal', internalInfix: 'int' }),
-    ).toStrictEqual([
-      {
-        name: 'default',
-        source: { url: 'https://raw.githubusercontent.com/org/repo/main/.readyup/kits/default.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'github:org/repo@main' },
-      },
-    ]);
-  });
-
-  it('ignores internal config when --url is used', () => {
-    expect(
-      resolve({
-        urlValue: 'https://example.com/config.js',
-        internal: true,
-        internalDir: 'internal',
-        internalInfix: 'int',
-      }),
-    ).toStrictEqual([
-      {
-        name: 'config',
-        source: { url: 'https://example.com/config.js' },
-        checklists: [],
-        provenance: { kind: 'remote', label: 'example.com/config.js' },
-      },
-    ]);
-  });
-});
+import { runCommand } from '../runCommand.ts';
 
 /**
  * The blank line parting one block from the next, as it reads in concatenated stdout writes.
