+++ packages/readyup/src/run/defaultKitName.ts
@@ -1,0 +1,2 @@
+/** The kit every source runs when the invocation names none. */
+export const DEFAULT_KIT_NAME = 'default';
