+++ packages/readyup/src/run/runCommand.ts
@@ -6,15 +6,10 @@ import { describeError } from '@williamthorsen/toolbelt.errors/candidate';
 import { EXIT_OK, EXIT_PROBLEMS_FOUND, EXIT_TOOL_FAILURE } from '../bin/exitCodes.ts';
 import { extractHint } from '../errors/error-handling.ts';
 import { kitLoadError, type RdyError, toRdyError, usageError } from '../errors/RdyError.ts';
-import { expandConfiguredPackages, type PackageKit } from '../installed-packages/expandConfiguredPackages.ts';
-import { readPackageVersion, resolvePackageRoot } from '../installed-packages/resolvePackageRoot.ts';
 import { describeUnresolvableImports } from '../kitImports/describeUnresolvableImports.ts';
 import { UnresolvableKitImportsError } from '../kitImports/UnresolvableKitImportsError.ts';
-import { buildKitFilename } from '../kits/buildKitFilename.ts';
 import type { KitProvenance } from '../kits/KitProvenance.ts';
-import { KITS_DIR, resolveHomeDir } from '../kits/kitsDir.ts';
 import { type LoadedRdyKit, loadRdyKit } from '../kits/loadRdyKit.ts';
-import { type FromSource, type NpmSource, parseFromValue } from '../kits/parseFromValue.ts';
 import type { FixLocation, RdyChecklist, RdyKit, RdyReport, RdyStagedChecklist, Severity } from '../kits/types.ts';
 import { getLayout } from '../layout/engine.ts';
 import type { BreadcrumbSegment, SummaryRow } from '../layout/layoutEngine.ts';
@@ -31,264 +26,9 @@ import type { JsonWarning, RaisedWarning } from '../schemas/common.ts';
 import type { JsonDetail, JsonKitOrigin } from '../schemas/reportSchema.ts';
 import { checkDrift } from '../verify/checkDrift.ts';
 import { checkSourceDrift } from '../verify/checkSourceDrift.ts';
-import type { KitSpecifier } from './parseKitSpecifiers.ts';
+import type { KitSource, ResolvedKitEntry } from './ResolvedKitEntry.ts';
 import { resolveRequestedNames } from './resolveRequestedNames.ts';
 import { runRdy } from './runRdy.ts';
-
-/** The kit every source runs when the invocation names none. */
-const DEFAULT_KIT_NAME = 'default';
-
-/** Discriminated union describing how to locate the rdy kit. */
-export type KitSource = { path: string } | { url: string };
-
-/** A resolved kit entry with its source and checklist filter. */
-export interface ResolvedKitEntry {
-  name: string;
-  source: KitSource;
-  checklists: string[];
-  provenance?: KitProvenance;
-}
-
-/** Build the GitHub raw content URL for a kit. */
-function buildGitHubKitUrl(org: string, repo: string, ref: string, kit: string, extension: string): string {
-  return `https://raw.githubusercontent.com/${org}/${repo}/${ref}/${KITS_DIR}/${kit}${extension}`;
-}
-
-/** Build the Bitbucket Cloud API source URL for a kit. */
-function buildBitbucketKitUrl(workspace: string, repo: string, ref: string, kit: string, extension: string): string {
-  return `https://api.bitbucket.org/2.0/repositories/${workspace}/${repo}/src/${ref}/${KITS_DIR}/${kit}${extension}`;
-}
-
-/** Resolve parsed flags into an array of kit entries to execute. */
-export function resolveKitSources({
-  filePath,
-  fromValue,
-  urlValue,
-  kitSpecifiers,
-  checklists,
-  jit,
-  internal,
-  internalDir,
-  internalInfix,
-  packages,
-  configuredPackages,
-}: {
-  filePath: string | undefined;
-  fromValue: string | undefined;
-  urlValue: string | undefined;
-  kitSpecifiers: KitSpecifier[];
-  checklists: string[] | undefined;
-  jit: boolean;
-  internal: boolean;
-  internalDir?: string | undefined;
-  internalInfix?: string | undefined;
-  packages?: boolean;
-  configuredPackages?: string[] | undefined;
-}): ResolvedKitEntry[] {
-  if (filePath !== undefined) {
-    return [
-      {
-        name: path.basename(filePath, path.extname(filePath)),
-        source: { path: filePath },
-        checklists: checklists ?? [],
-        provenance: { kind: 'directory', label: path.dirname(filePath) },
-      },
-    ];
-  }
-  if (urlValue !== undefined) {
-    const { label, name } = describeUrlSource(urlValue);
-    return [{ name, source: { url: urlValue }, checklists: checklists ?? [], provenance: { kind: 'remote', label } }];
-  }
-
-  // Assume `jit` is always false when `fromValue` is present; `parseRunArgs` enforces this constraint.
-  const extension = jit ? '.ts' : '.js';
-
-  // Fill the default before the `--packages` branch reads it, so a bare invocation is structurally
-  // `--packages default` and the two forms cannot select different kits.
-  const declaredSpecs = kitSpecifiers.length > 0 ? kitSpecifiers : [{ kitName: DEFAULT_KIT_NAME, checklists: [] }];
-
-  if (packages === true) {
-    const requestedNames = declaredSpecs.map((spec) => spec.kitName);
-    return resolveConfiguredPackages(configuredPackages ?? [], requestedNames, extension);
-  }
-
-  // `--checklists` names checklists within one kit, and `parseRunArgs` has already rejected every
-  // invocation where "one kit" is ambiguous, so this map never covers more than a single spec.
-  const specs = checklists === undefined ? declaredSpecs : declaredSpecs.map((spec) => ({ ...spec, checklists }));
-
-  if (fromValue !== undefined) {
-    let source: FromSource;
-    try {
-      source = parseFromValue(fromValue);
-    } catch (error: unknown) {
-      throw usageError(describeError(error), { cause: error });
-    }
-    return resolveFromSource(source, specs, extension);
-  }
-
-  // Default/internal case: resolve from the current repo.
-  if (internal) {
-    return specs.map((spec) => ({
-      name: spec.kitName,
-      source: {
-        path: path.join(KITS_DIR, internalDir ?? '.', buildKitFilename(spec.kitName, internalInfix, extension)),
-      },
-      checklists: spec.checklists,
-    }));
-  }
-
-  return specs.map((spec) => ({
-    name: spec.kitName,
-    source: { path: path.join(KITS_DIR, `${spec.kitName}${extension}`) },
-    checklists: spec.checklists,
-  }));
-}
-
-/** Resolve kit entries from a parsed `--from` source. */
-function resolveFromSource(source: FromSource, specs: KitSpecifier[], extension: string): ResolvedKitEntry[] {
-  switch (source.type) {
-    case 'github': {
-      const provenance: KitProvenance = { kind: 'remote', label: `github:${source.org}/${source.repo}@${source.ref}` };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { url: buildGitHubKitUrl(source.org, source.repo, source.ref, spec.kitName, extension) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-
-    case 'bitbucket': {
-      const label = `bitbucket:${source.workspace}/${source.repo}@${source.ref}`;
-      const provenance: KitProvenance = { kind: 'remote', label };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { url: buildBitbucketKitUrl(source.workspace, source.repo, source.ref, spec.kitName, extension) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-
-    case 'npm': {
-      const root = resolveInstalledPackageRoot(source);
-      const provenance: KitProvenance = {
-        kind: 'package',
-        packageName: source.name,
-        version: readPackageVersion(root),
-      };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { path: path.join(root, KITS_DIR, `${spec.kitName}${extension}`) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-
-    case 'global': {
-      const homeDir = resolveHomeDir();
-      const provenance: KitProvenance = { kind: 'directory', label: `~/${KITS_DIR}` };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { path: path.join(homeDir, KITS_DIR, `${spec.kitName}${extension}`) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-
-    case 'directory': {
-      const provenance: KitProvenance = { kind: 'directory', label: source.path };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { path: path.join(path.resolve(process.cwd(), source.path), `${spec.kitName}${extension}`) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-
-    case 'local': {
-      const resolvedBase = path.resolve(process.cwd(), source.path);
-      const provenance: KitProvenance = { kind: 'directory', label: path.join(source.path, KITS_DIR) };
-      return specs.map((spec) => ({
-        name: spec.kitName,
-        source: { path: path.join(resolvedBase, KITS_DIR, `${spec.kitName}${extension}`) },
-        checklists: spec.checklists,
-        provenance,
-      }));
-    }
-  }
-}
-
-/**
- * Resolves the requested kits, drawn from what the configured packages publish, into run entries.
- *
- * An empty `packages` list is a usage error: the flag names a config key the config does not have, so
- * the invocation asks for something that cannot be answered. Configured packages that publish no
- * requested kit are a different case and run nothing, which is the honest answer to "does this project
- * satisfy what these packages require of it" when they require nothing.
- */
-function resolveConfiguredPackages(
-  configuredPackages: string[],
-  requestedNames: string[],
-  extension: string,
-): ResolvedKitEntry[] {
-  if (configuredPackages.length === 0) {
-    throw usageError('--packages requires a "packages" list in the readyup config; none is configured.');
-  }
-
-  const published = expandConfiguredPackages(configuredPackages, extension);
-
-  return selectRequestedKits(published, requestedNames).map((kit) => ({
-    name: kit.kitName,
-    source: { path: kit.path },
-    checklists: [],
-    provenance: { kind: 'package', packageName: kit.packageName, version: kit.version },
-  }));
-}
-
-/**
- * Narrows what the configured packages publish to the requested kits, name-major.
- *
- * A configured package not publishing a requested kit is skipped rather than reported: `--packages`
- * asks whether this project satisfies what its configured packages require of it, and a package
- * requiring nothing under that name has nothing to answer for.
- *
- * Name-major so `--packages a b` runs every package's `a` before any package's `b`, matching the
- * order `rdy run a b` runs them in against a single source.
- */
-function selectRequestedKits(published: PackageKit[], requestedNames: string[]): PackageKit[] {
-  return requestedNames.flatMap((kitName) => {
-    const selected = published.filter((kit) => kit.kitName === kitName);
-    if (selected.length === 0 && kitName !== DEFAULT_KIT_NAME) {
-      const available = [...new Set(published.map((kit) => kit.kitName))].join(', ');
-      throw usageError(`No configured package publishes a kit named "${kitName}"; available kits: ${available}.`);
-    }
-    return selected;
-  });
-}
-
-/**
- * Locates the root of a package named by `npm:`, rejecting the forms that are reserved but not yet served.
- *
- * A version spec is parsed rather than ignored so the syntax stays reserved for running a published
- * version; until that lands, naming one is answered with the flag that reaches a published kit today.
- *
- * The not-installed message names the direct-dependency requirement because pnpm's layout links only
- * direct dependencies into a project's `node_modules`. A transitive dependency is genuinely unreachable
- * here, and a bare "not installed" would contradict the lockfile the reader is looking at.
- */
-function resolveInstalledPackageRoot(source: NpmSource): string {
-  if (source.versionSpec !== undefined) {
-    throw usageError(
-      `Running a published version is not supported yet: "npm:${source.name}@${source.versionSpec}". ` +
-        'Use --url to name a published kit, or drop the version to run the installed copy.',
-    );
-  }
-
-  const root = resolvePackageRoot(source.name);
-  if (root === undefined) {
-    throw kitLoadError(`Package "${source.name}" is not installed; it must be a direct dependency of this project.`);
-  }
-  return root;
-}
 
 /** Resolve the effective fixLocation for a checklist, falling back to the kit-level default. */
 function resolveFixLocation(checklist: RdyChecklist | RdyStagedChecklist, kitDefault?: FixLocation): FixLocation {
@@ -766,20 +506,6 @@ function describeKitProvenance(provenance: KitProvenance | undefined): Breadcrum
 
   const version = provenance.version === undefined ? '' : `@${provenance.version}`;
   return { role: 'sourcePackage', text: `${provenance.packageName}${version}` };
-}
-
-/**
- * Splits a kit URL into the kit's name and the label naming where it was fetched from.
- *
- * The scheme is dropped from the label because every kit URL carries one and it distinguishes nothing.
- * A URL that does not parse is reported exactly as given, since a value the runner could not read is one
- * the reader needs to see unaltered.
- */
-function describeUrlSource(urlValue: string): { label: string; name: string } {
-  if (!URL.canParse(urlValue)) return { label: urlValue, name: urlValue };
-
-  const { host, pathname } = new URL(urlValue);
-  return { label: `${host}${pathname}`, name: path.basename(pathname, path.extname(pathname)) };
 }
 
 /** Resolves a kit's requested checklist names to the checklists themselves, in requested order. */
